fix: client args are now parse in the client function
